backend/requirements.txt
fastapi
uvicorn==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4
